% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcwordsimilarityfunction.R
\name{calculate_word_similarity}
\alias{calculate_word_similarity}
\title{Calculate Word Similarity within a chRoma Database}
\usage{
calculate_word_similarity(chRoma_db, word1, word2)
}
\arguments{
\item{chRoma_db}{The chRoma database where word vectors are stored.}

\item{word1}{The first word.}

\item{word2}{The second word.}
}
\value{
The similarity score between the two words.
}
\description{
Calculate the similarity between two words based on their word vectors within a chRoma database.
}
\examples{
db <- create_collection()
word_vectors <- data.frame(word = c("apple", "banana", "cherry"),
                           vector = list(c(0.2, 0.4, 0.6), c(0.1, 0.3, 0.5), c(0.4, 0.2, 0.8)))
add_collection(db, word_vectors)
word1 <- "apple"
word2 <- "banana"
calculate_word_similarity(db, word1, word2)

}
